--- pages/main_page/MainPage.js
@@ -1,16 +1,0 @@
-import React, { useRef } from 'react';
-import Intro from '../../components/sections/Intro';
-import AboutMe from '../../components/sections/Aboutme';
-
-function MainPage() {
-  const aboutMeRef = useRef(null);
-
-  return (
-    <>
-      <Intro aboutMeRef={aboutMeRef} />
-      <AboutMe ref={aboutMeRef} />
-    </>
-  );
-}
-
-export default MainPage;
